--- src/components/common/Modal.tsx
@@ -1,3 +1,0 @@
-export default function Modal() {
-  return <div>a</div>;
-}
